+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+* Fixed an issue were doc_images required release-notes and validations.
+
+# 1.2.0
 * Fixed an issue where **format** did not update the test playbook from its pack.
 * Fixed an issue where **validate** validated non integration images.
 * Fixed an issue where **update-release-notes** did not identified old yml integrations and scripts.
@@ -13,6 +16,7 @@
 * Fixed an issue where image and description files were not handled correctly by **validate** and **update-release-notes** commands.
 * Fixed an issue where the **format** command didn't remove all extra fields in a file.
 * Added an error in case an invalid id_set.json file is found while running the **validate** command.
+* Added fetch params checks to the **validate** command.
 
 # 1.1.11
 * Added line number to secrets' path in **secrets** command report.